requirements.txt
google-cloud-aiplatform[adk,agent_engines]
requests
python-dotenv
# chainlit
streamlit